feat(mirror): guest emits CMD_SCENE_REGISTER per static object (Phase 1)
Behind flametal:native_scene_mirror (default off), drawEntryOnGpu's retained
branch now emits one SCENE_REGISTER per static object: {objectId=scene ptr,
meshId=retained.meshId, signature (describe content hash), vertexCount,
materialFlags, world=identity}. The guest does NOT skip its draw-walk --
log-only/observational; production behaviour byte-identical with the flag off.

Widened the protocol signature field uint32 -> uint64 (describe yields a
64-bit content hash; the mirror needs the full width for its eventual
content-invalidation). Struct 96 -> 100 bytes; re-verified identical layout
on x86_64 (guest) and arm64 (host). sceneRegister API param widened to match.

Bounds (center/radius) are not in drawEntryOnGpu's scope (they live on the
caller's pObj57AD20 vec/f20) -- emitted as zero with a TODO; threaded in a
later phase when the host actually culls. SCENE_RESET level-load wiring
deferred to the validation iteration.

+++ src/gog_patch_dll/metal_bridge/MetalBridgeProducer.cpp
@@ -2172,7 +2172,7 @@ public:
 
     // Native scene mirror (Phase 1, log-only). Stages a per-object register or
     // an epoch bump; the host builds a registry and consumes nothing yet.
-    void sceneRegister(uint32_t objectId, uint32_t meshId, uint32_t signature,
+    void sceneRegister(uint32_t objectId, uint32_t meshId, uint64_t signature,
                        uint32_t vertexCount, uint32_t materialFlags,
                        const float world[12], const float center[3], float radius) {
         if (!active_) return;  // mirror is host-only; nothing to stage without one
